Updated to use WIN+Mousebuttons + horizontal-only
Some people notified me that they prefer to use the middle mouse button for the popup that Studio One gives so they can select a different tool. The same functionality for dragging could also be achieved by simply switching to a WIN+LeftMousebutton type of activation, so I changed it to that.